fix: read application via request from `AppController::initialize`
