+++ plugins/BEdita/API/src/Controller/AppController.php
@@ -12,10 +12,14 @@
  */
 namespace BEdita\API\Controller;
 
+use BadMethodCallException;
 use BEdita\API\Datasource\JsonApiPaginator;
+use BEdita\Core\State\CurrentApplication;
 use Cake\Controller\Controller;
 use Cake\Core\Configure;
+use Cake\Datasource\Exception\RecordNotFoundException;
 use Cake\Event\Event;
+use Cake\Network\Exception\ForbiddenException;
 use Cake\Network\Exception\NotAcceptableException;
 use Cake\Routing\Router;
 
@@ -48,6 +52,8 @@ class AppController extends Controller
 
         $this->response = $this->response->withHeader('X-BEdita-Version', Configure::read('BEdita.version'));
 
+        $this->getApplication();
+
         $this->loadComponent('Paginator', (array)Configure::read('Pagination'));
         $this->loadComponent('RequestHandler');
         if ($this->request->is(['json', 'jsonapi'])) {
@@ -65,7 +71,6 @@ class AppController extends Controller
             'authenticate' => ['BEdita/API.Jwt', 'BEdita/API.Anonymous'],
             'authorize' => [
                 'BEdita/API.Endpoint' => [
-                    'blockAnonymousApps' => Configure::read('Security.blockAnonymousApps'),
                     'blockAnonymousUsers' => Configure::read('Security.blockAnonymousUsers'),
                 ],
             ],
@@ -97,5 +102,34 @@ class AppController extends Controller
         }
 
         return null;
+    }
+
+    /**
+     * Get application from request.
+     * This is done primarily with an API_KEY header like 'X-Api-Key',
+     * alternatively `api_key` query string is used (not recommended)
+     *
+     * @return void
+     * @throws \Cake\Network\Exception\ForbiddenException Throws an exception if API key is missing or invalid.
+     */
+    protected function getApplication()
+    {
+        if (CurrentApplication::getApplication() === null) {
+            $apiKey = $this->request->getHeaderLine('X-Api-Key');
+            if (empty($apiKey)) {
+                $apiKey = (string)$this->request->getQuery('api_key');
+            }
+            if (empty($apiKey) && empty(Configure::read('Security.blockAnonymousApps'))) {
+                return;
+            }
+
+            try {
+                CurrentApplication::setFromApiKey($apiKey);
+            } catch (BadMethodCallException $e) {
+                throw new ForbiddenException(__d('bedita', 'Missing API key'));
+            } catch (RecordNotFoundException $e) {
+                throw new ForbiddenException(__d('bedita', 'Invalid API key'));
+            }
+        }
     }
 }
